feat(cache): attribute staged planner outcomes
